+++ functions/src/models/index.ts
@@ -4,3 +4,4 @@ export * from "./EntityPair";
 export * from "./Delivery";
 export * from "./DodoToken";
 export * from "./DodoOrder";
+export * from "./FoodBoxesCheckupReportedCount";