+++ lib/common/presentation/router/app_router.gr.dart
@@ -9,61 +9,65 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i22;
-import 'package:collection/collection.dart' as _i25;
-import 'package:flutter/material.dart' as _i23;
-import 'package:zachranobed/app/presentation/home_screen.dart' as _i13;
-import 'package:zachranobed/common/domain/model/user_data.dart' as _i26;
+import 'package:auto_route/auto_route.dart' as _i23;
+import 'package:collection/collection.dart' as _i26;
+import 'package:flutter/material.dart' as _i24;
+import 'package:zachranobed/app/presentation/home_screen.dart' as _i14;
+import 'package:zachranobed/common/domain/model/user_data.dart' as _i27;
 import 'package:zachranobed/features/activepair/presentation/change_active_pair_screen.dart'
     as _i2;
 import 'package:zachranobed/features/appTerms/presentation/app_terms_screen.dart'
     as _i1;
 import 'package:zachranobed/features/debug/components_screen.dart' as _i4;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i6;
-import 'package:zachranobed/features/faq/domain/model/faq_item.dart' as _i24;
+import 'package:zachranobed/features/faq/domain/model/faq_item.dart' as _i25;
 import 'package:zachranobed/features/faq/presentation/screen/faq_detail_screen.dart'
     as _i8;
 import 'package:zachranobed/features/faq/presentation/screen/faq_questions_screen.dart'
     as _i9;
+import 'package:zachranobed/features/food/domain/model/food_box_statistics.dart'
+    as _i28;
 import 'package:zachranobed/features/food/domain/model/food_box_type.dart'
-    as _i27;
-import 'package:zachranobed/features/food/domain/model/food_info.dart' as _i28;
+    as _i29;
+import 'package:zachranobed/features/food/domain/model/food_info.dart' as _i30;
 import 'package:zachranobed/features/food/presentation/screens/delivery_detail_screen.dart'
     as _i7;
-import 'package:zachranobed/features/food/presentation/screens/food_boxes_detail_screen.dart'
+import 'package:zachranobed/features/food/presentation/screens/food_boxes_checkup_mismatch_screen.dart'
     as _i10;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_boxes_screen.dart'
-    as _i16;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_detail_screen.dart'
-    as _i15;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_overview_screen.dart'
-    as _i17;
-import 'package:zachranobed/features/food/presentation/screens/order_shipping_of_boxes_screen.dart'
-    as _i18;
-import 'package:zachranobed/features/food/presentation/screens/thank_you_screen.dart'
-    as _i20;
-import 'package:zachranobed/features/forceupdate/presentation/force_update_screen.dart'
+import 'package:zachranobed/features/food/presentation/screens/food_boxes_detail_screen.dart'
     as _i11;
+import 'package:zachranobed/features/food/presentation/screens/offer_food_boxes_screen.dart'
+    as _i17;
+import 'package:zachranobed/features/food/presentation/screens/offer_food_detail_screen.dart'
+    as _i16;
+import 'package:zachranobed/features/food/presentation/screens/offer_food_overview_screen.dart'
+    as _i18;
+import 'package:zachranobed/features/food/presentation/screens/order_shipping_of_boxes_screen.dart'
+    as _i19;
+import 'package:zachranobed/features/food/presentation/screens/thank_you_screen.dart'
+    as _i21;
+import 'package:zachranobed/features/forceupdate/presentation/force_update_screen.dart'
+    as _i12;
 import 'package:zachranobed/features/login/presentation/screen/change_password_screen.dart'
     as _i3;
 import 'package:zachranobed/features/login/presentation/screen/forgot_password_screen.dart'
-    as _i12;
+    as _i13;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
-    as _i14;
+    as _i15;
 import 'package:zachranobed/features/menu/presentation/contacts_screen.dart'
     as _i5;
 import 'package:zachranobed/features/menu/presentation/profile_screen.dart'
-    as _i19;
+    as _i20;
 import 'package:zachranobed/features/whatsnew/presentation/whats_new_screen.dart'
-    as _i21;
+    as _i22;
 
 /// generated route for
 /// [_i1.AppTermsScreen]
-class AppTermsRoute extends _i22.PageRouteInfo<AppTermsRouteArgs> {
+class AppTermsRoute extends _i23.PageRouteInfo<AppTermsRouteArgs> {
   AppTermsRoute({
-    _i23.Key? key,
+    _i24.Key? key,
     required bool hasNoAcceptedVersion,
-    List<_i22.PageRouteInfo>? children,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           AppTermsRoute.name,
           args: AppTermsRouteArgs(
@@ -75,7 +79,7 @@ class AppTermsRoute extends _i22.PageRouteInfo<AppTermsRouteArgs> {
 
   static const String name = 'AppTermsRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<AppTermsRouteArgs>();
@@ -90,7 +94,7 @@ class AppTermsRoute extends _i22.PageRouteInfo<AppTermsRouteArgs> {
 class AppTermsRouteArgs {
   const AppTermsRouteArgs({this.key, required this.hasNoAcceptedVersion});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
   final bool hasNoAcceptedVersion;
 
@@ -113,13 +117,13 @@ class AppTermsRouteArgs {
 
 /// generated route for
 /// [_i2.ChangeActivePairScreen]
-class ChangeActivePairRoute extends _i22.PageRouteInfo<void> {
-  const ChangeActivePairRoute({List<_i22.PageRouteInfo>? children})
+class ChangeActivePairRoute extends _i23.PageRouteInfo<void> {
+  const ChangeActivePairRoute({List<_i23.PageRouteInfo>? children})
       : super(ChangeActivePairRoute.name, initialChildren: children);
 
   static const String name = 'ChangeActivePairRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       return const _i2.ChangeActivePairScreen();
@@ -129,13 +133,13 @@ class ChangeActivePairRoute extends _i22.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i3.ChangePasswordScreen]
-class ChangePasswordRoute extends _i22.PageRouteInfo<void> {
-  const ChangePasswordRoute({List<_i22.PageRouteInfo>? children})
+class ChangePasswordRoute extends _i23.PageRouteInfo<void> {
+  const ChangePasswordRoute({List<_i23.PageRouteInfo>? children})
       : super(ChangePasswordRoute.name, initialChildren: children);
 
   static const String name = 'ChangePasswordRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       return const _i3.ChangePasswordScreen();
@@ -145,13 +149,13 @@ class ChangePasswordRoute extends _i22.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i4.ComponentsScreen]
-class ComponentsRoute extends _i22.PageRouteInfo<void> {
-  const ComponentsRoute({List<_i22.PageRouteInfo>? children})
+class ComponentsRoute extends _i23.PageRouteInfo<void> {
+  const ComponentsRoute({List<_i23.PageRouteInfo>? children})
       : super(ComponentsRoute.name, initialChildren: children);
 
   static const String name = 'ComponentsRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       return const _i4.ComponentsScreen();
@@ -161,13 +165,13 @@ class ComponentsRoute extends _i22.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i5.ContactsScreen]
-class ContactsRoute extends _i22.PageRouteInfo<void> {
-  const ContactsRoute({List<_i22.PageRouteInfo>? children})
+class ContactsRoute extends _i23.PageRouteInfo<void> {
+  const ContactsRoute({List<_i23.PageRouteInfo>? children})
       : super(ContactsRoute.name, initialChildren: children);
 
   static const String name = 'ContactsRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       return const _i5.ContactsScreen();
@@ -177,13 +181,13 @@ class ContactsRoute extends _i22.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i6.DebugScreen]
-class DebugRoute extends _i22.PageRouteInfo<void> {
-  const DebugRoute({List<_i22.PageRouteInfo>? children})
+class DebugRoute extends _i23.PageRouteInfo<void> {
+  const DebugRoute({List<_i23.PageRouteInfo>? children})
       : super(DebugRoute.name, initialChildren: children);
 
   static const String name = 'DebugRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       return const _i6.DebugScreen();
@@ -193,11 +197,11 @@ class DebugRoute extends _i22.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i7.DeliveryDetailScreen]
-class DeliveryDetailRoute extends _i22.PageRouteInfo<DeliveryDetailRouteArgs> {
+class DeliveryDetailRoute extends _i23.PageRouteInfo<DeliveryDetailRouteArgs> {
   DeliveryDetailRoute({
-    _i23.Key? key,
+    _i24.Key? key,
     required String deliveryId,
-    List<_i22.PageRouteInfo>? children,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           DeliveryDetailRoute.name,
           args: DeliveryDetailRouteArgs(key: key, deliveryId: deliveryId),
@@ -206,7 +210,7 @@ class DeliveryDetailRoute extends _i22.PageRouteInfo<DeliveryDetailRouteArgs> {
 
   static const String name = 'DeliveryDetailRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<DeliveryDetailRouteArgs>();
@@ -221,7 +225,7 @@ class DeliveryDetailRoute extends _i22.PageRouteInfo<DeliveryDetailRouteArgs> {
 class DeliveryDetailRouteArgs {
   const DeliveryDetailRouteArgs({this.key, required this.deliveryId});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
   final String deliveryId;
 
@@ -243,11 +247,11 @@ class DeliveryDetailRouteArgs {
 
 /// generated route for
 /// [_i8.FaqDetailScreen]
-class FaqDetailRoute extends _i22.PageRouteInfo<FaqDetailRouteArgs> {
+class FaqDetailRoute extends _i23.PageRouteInfo<FaqDetailRouteArgs> {
   FaqDetailRoute({
-    _i23.Key? key,
-    required _i24.FaqItem item,
-    List<_i22.PageRouteInfo>? children,
+    _i24.Key? key,
+    required _i25.FaqItem item,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           FaqDetailRoute.name,
           args: FaqDetailRouteArgs(key: key, item: item),
@@ -256,7 +260,7 @@ class FaqDetailRoute extends _i22.PageRouteInfo<FaqDetailRouteArgs> {
 
   static const String name = 'FaqDetailRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FaqDetailRouteArgs>();
@@ -268,9 +272,9 @@ class FaqDetailRoute extends _i22.PageRouteInfo<FaqDetailRouteArgs> {
 class FaqDetailRouteArgs {
   const FaqDetailRouteArgs({this.key, required this.item});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final _i24.FaqItem item;
+  final _i25.FaqItem item;
 
   @override
   String toString() {
@@ -290,11 +294,11 @@ class FaqDetailRouteArgs {
 
 /// generated route for
 /// [_i9.FaqQuestionsScreen]
-class FaqQuestionsRoute extends _i22.PageRouteInfo<FaqQuestionsRouteArgs> {
+class FaqQuestionsRoute extends _i23.PageRouteInfo<FaqQuestionsRouteArgs> {
   FaqQuestionsRoute({
-    _i23.Key? key,
-    required List<_i24.FaqItem> items,
-    List<_i22.PageRouteInfo>? children,
+    _i24.Key? key,
+    required List<_i25.FaqItem> items,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           FaqQuestionsRoute.name,
           args: FaqQuestionsRouteArgs(key: key, items: items),
@@ -303,7 +307,7 @@ class FaqQuestionsRoute extends _i22.PageRouteInfo<FaqQuestionsRouteArgs> {
 
   static const String name = 'FaqQuestionsRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FaqQuestionsRouteArgs>();
@@ -315,9 +319,9 @@ class FaqQuestionsRoute extends _i22.PageRouteInfo<FaqQuestionsRouteArgs> {
 class FaqQuestionsRouteArgs {
   const FaqQuestionsRouteArgs({this.key, required this.items});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final List<_i24.FaqItem> items;
+  final List<_i25.FaqItem> items;
 
   @override
   String toString() {
@@ -329,22 +333,88 @@ class FaqQuestionsRouteArgs {
     if (identical(this, other)) return true;
     if (other is! FaqQuestionsRouteArgs) return false;
     return key == other.key &&
-        const _i25.ListEquality().equals(items, other.items);
+        const _i26.ListEquality().equals(items, other.items);
   }
 
   @override
-  int get hashCode => key.hashCode ^ const _i25.ListEquality().hash(items);
+  int get hashCode => key.hashCode ^ const _i26.ListEquality().hash(items);
 }
 
 /// generated route for
-/// [_i10.FoodBoxesDetailScreen]
+/// [_i10.FoodBoxesCheckupMismatchScreen]
+class FoodBoxesCheckupMismatchRoute
+    extends _i23.PageRouteInfo<FoodBoxesCheckupMismatchRouteArgs> {
+  FoodBoxesCheckupMismatchRoute({
+    _i24.Key? key,
+    required _i27.UserData user,
+    required List<_i28.FoodBoxStatistics> statistics,
+    List<_i23.PageRouteInfo>? children,
+  }) : super(
+          FoodBoxesCheckupMismatchRoute.name,
+          args: FoodBoxesCheckupMismatchRouteArgs(
+            key: key,
+            user: user,
+            statistics: statistics,
+          ),
+          initialChildren: children,
+        );
+
+  static const String name = 'FoodBoxesCheckupMismatchRoute';
+
+  static _i23.PageInfo page = _i23.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<FoodBoxesCheckupMismatchRouteArgs>();
+      return _i10.FoodBoxesCheckupMismatchScreen(
+        key: args.key,
+        user: args.user,
+        statistics: args.statistics,
+      );
+    },
+  );
+}
+
+class FoodBoxesCheckupMismatchRouteArgs {
+  const FoodBoxesCheckupMismatchRouteArgs({
+    this.key,
+    required this.user,
+    required this.statistics,
+  });
+
+  final _i24.Key? key;
+
+  final _i27.UserData user;
+
+  final List<_i28.FoodBoxStatistics> statistics;
+
+  @override
+  String toString() {
+    return 'FoodBoxesCheckupMismatchRouteArgs{key: $key, user: $user, statistics: $statistics}';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is! FoodBoxesCheckupMismatchRouteArgs) return false;
+    return key == other.key &&
+        user == other.user &&
+        const _i26.ListEquality().equals(statistics, other.statistics);
+  }
+
+  @override
+  int get hashCode =>
+      key.hashCode ^ user.hashCode ^ const _i26.ListEquality().hash(statistics);
+}
+
+/// generated route for
+/// [_i11.FoodBoxesDetailScreen]
 class FoodBoxesDetailRoute
-    extends _i22.PageRouteInfo<FoodBoxesDetailRouteArgs> {
+    extends _i23.PageRouteInfo<FoodBoxesDetailRouteArgs> {
   FoodBoxesDetailRoute({
-    _i23.Key? key,
-    required _i26.UserData user,
+    _i24.Key? key,
+    required _i27.UserData user,
     bool isCheckupMode = false,
-    List<_i22.PageRouteInfo>? children,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           FoodBoxesDetailRoute.name,
           args: FoodBoxesDetailRouteArgs(
@@ -357,11 +427,11 @@ class FoodBoxesDetailRoute
 
   static const String name = 'FoodBoxesDetailRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FoodBoxesDetailRouteArgs>();
-      return _i10.FoodBoxesDetailScreen(
+      return _i11.FoodBoxesDetailScreen(
         key: args.key,
         user: args.user,
         isCheckupMode: args.isCheckupMode,
@@ -377,9 +447,9 @@ class FoodBoxesDetailRouteArgs {
     this.isCheckupMode = false,
   });
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final _i26.UserData user;
+  final _i27.UserData user;
 
   final bool isCheckupMode;
 
@@ -402,44 +472,44 @@ class FoodBoxesDetailRouteArgs {
 }
 
 /// generated route for
-/// [_i11.ForceUpdateScreen]
-class ForceUpdateRoute extends _i22.PageRouteInfo<void> {
-  const ForceUpdateRoute({List<_i22.PageRouteInfo>? children})
+/// [_i12.ForceUpdateScreen]
+class ForceUpdateRoute extends _i23.PageRouteInfo<void> {
+  const ForceUpdateRoute({List<_i23.PageRouteInfo>? children})
       : super(ForceUpdateRoute.name, initialChildren: children);
 
   static const String name = 'ForceUpdateRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i11.ForceUpdateScreen();
+      return const _i12.ForceUpdateScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i12.ForgotPasswordScreen]
-class ForgotPasswordRoute extends _i22.PageRouteInfo<void> {
-  const ForgotPasswordRoute({List<_i22.PageRouteInfo>? children})
+/// [_i13.ForgotPasswordScreen]
+class ForgotPasswordRoute extends _i23.PageRouteInfo<void> {
+  const ForgotPasswordRoute({List<_i23.PageRouteInfo>? children})
       : super(ForgotPasswordRoute.name, initialChildren: children);
 
   static const String name = 'ForgotPasswordRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i12.ForgotPasswordScreen();
+      return const _i13.ForgotPasswordScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i13.HomeScreen]
-class HomeRoute extends _i22.PageRouteInfo<HomeRouteArgs> {
+/// [_i14.HomeScreen]
+class HomeRoute extends _i23.PageRouteInfo<HomeRouteArgs> {
   HomeRoute({
-    _i23.Key? key,
+    _i24.Key? key,
     int initialTabIndex = 0,
-    List<_i22.PageRouteInfo>? children,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           HomeRoute.name,
           args: HomeRouteArgs(key: key, initialTabIndex: initialTabIndex),
@@ -448,13 +518,13 @@ class HomeRoute extends _i22.PageRouteInfo<HomeRouteArgs> {
 
   static const String name = 'HomeRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<HomeRouteArgs>(
         orElse: () => const HomeRouteArgs(),
       );
-      return _i13.HomeScreen(
+      return _i14.HomeScreen(
         key: args.key,
         initialTabIndex: args.initialTabIndex,
       );
@@ -465,7 +535,7 @@ class HomeRoute extends _i22.PageRouteInfo<HomeRouteArgs> {
 class HomeRouteArgs {
   const HomeRouteArgs({this.key, this.initialTabIndex = 0});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
   final int initialTabIndex;
 
@@ -486,44 +556,44 @@ class HomeRouteArgs {
 }
 
 /// generated route for
-/// [_i14.LoginScreen]
-class LoginRoute extends _i22.PageRouteInfo<void> {
-  const LoginRoute({List<_i22.PageRouteInfo>? children})
+/// [_i15.LoginScreen]
+class LoginRoute extends _i23.PageRouteInfo<void> {
+  const LoginRoute({List<_i23.PageRouteInfo>? children})
       : super(LoginRoute.name, initialChildren: children);
 
   static const String name = 'LoginRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i14.LoginScreen();
+      return const _i15.LoginScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i15.OfferFoodAddNewScreen]
-class OfferFoodAddNewRoute extends _i22.PageRouteInfo<void> {
-  const OfferFoodAddNewRoute({List<_i22.PageRouteInfo>? children})
+/// [_i16.OfferFoodAddNewScreen]
+class OfferFoodAddNewRoute extends _i23.PageRouteInfo<void> {
+  const OfferFoodAddNewRoute({List<_i23.PageRouteInfo>? children})
       : super(OfferFoodAddNewRoute.name, initialChildren: children);
 
   static const String name = 'OfferFoodAddNewRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i15.OfferFoodAddNewScreen();
+      return const _i16.OfferFoodAddNewScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i16.OfferFoodBoxesScreen]
-class OfferFoodBoxesRoute extends _i22.PageRouteInfo<OfferFoodBoxesRouteArgs> {
+/// [_i17.OfferFoodBoxesScreen]
+class OfferFoodBoxesRoute extends _i23.PageRouteInfo<OfferFoodBoxesRouteArgs> {
   OfferFoodBoxesRoute({
-    _i23.Key? key,
-    required Map<_i27.FoodBoxType, int> currentBoxesQuantity,
-    List<_i22.PageRouteInfo>? children,
+    _i24.Key? key,
+    required Map<_i29.FoodBoxType, int> currentBoxesQuantity,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           OfferFoodBoxesRoute.name,
           args: OfferFoodBoxesRouteArgs(
@@ -535,11 +605,11 @@ class OfferFoodBoxesRoute extends _i22.PageRouteInfo<OfferFoodBoxesRouteArgs> {
 
   static const String name = 'OfferFoodBoxesRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodBoxesRouteArgs>();
-      return _i16.OfferFoodBoxesScreen(
+      return _i17.OfferFoodBoxesScreen(
         key: args.key,
         currentBoxesQuantity: args.currentBoxesQuantity,
       );
@@ -550,9 +620,9 @@ class OfferFoodBoxesRoute extends _i22.PageRouteInfo<OfferFoodBoxesRouteArgs> {
 class OfferFoodBoxesRouteArgs {
   const OfferFoodBoxesRouteArgs({this.key, required this.currentBoxesQuantity});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final Map<_i27.FoodBoxType, int> currentBoxesQuantity;
+  final Map<_i29.FoodBoxType, int> currentBoxesQuantity;
 
   @override
   String toString() {
@@ -564,7 +634,7 @@ class OfferFoodBoxesRouteArgs {
     if (identical(this, other)) return true;
     if (other is! OfferFoodBoxesRouteArgs) return false;
     return key == other.key &&
-        const _i25.MapEquality().equals(
+        const _i26.MapEquality().equals(
           currentBoxesQuantity,
           other.currentBoxesQuantity,
         );
@@ -572,17 +642,17 @@ class OfferFoodBoxesRouteArgs {
 
   @override
   int get hashCode =>
-      key.hashCode ^ const _i25.MapEquality().hash(currentBoxesQuantity);
+      key.hashCode ^ const _i26.MapEquality().hash(currentBoxesQuantity);
 }
 
 /// generated route for
-/// [_i15.OfferFoodEditExistingScreen]
+/// [_i16.OfferFoodEditExistingScreen]
 class OfferFoodEditExistingRoute
-    extends _i22.PageRouteInfo<OfferFoodEditExistingRouteArgs> {
+    extends _i23.PageRouteInfo<OfferFoodEditExistingRouteArgs> {
   OfferFoodEditExistingRoute({
-    _i23.Key? key,
-    required _i28.FoodInfo foodInfo,
-    List<_i22.PageRouteInfo>? children,
+    _i24.Key? key,
+    required _i30.FoodInfo foodInfo,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           OfferFoodEditExistingRoute.name,
           args: OfferFoodEditExistingRouteArgs(key: key, foodInfo: foodInfo),
@@ -591,11 +661,11 @@ class OfferFoodEditExistingRoute
 
   static const String name = 'OfferFoodEditExistingRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodEditExistingRouteArgs>();
-      return _i15.OfferFoodEditExistingScreen(
+      return _i16.OfferFoodEditExistingScreen(
         key: args.key,
         foodInfo: args.foodInfo,
       );
@@ -606,9 +676,9 @@ class OfferFoodEditExistingRoute
 class OfferFoodEditExistingRouteArgs {
   const OfferFoodEditExistingRouteArgs({this.key, required this.foodInfo});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final _i28.FoodInfo foodInfo;
+  final _i30.FoodInfo foodInfo;
 
   @override
   String toString() {
@@ -627,29 +697,29 @@ class OfferFoodEditExistingRouteArgs {
 }
 
 /// generated route for
-/// [_i15.OfferFoodInitialScreen]
-class OfferFoodInitialRoute extends _i22.PageRouteInfo<void> {
-  const OfferFoodInitialRoute({List<_i22.PageRouteInfo>? children})
+/// [_i16.OfferFoodInitialScreen]
+class OfferFoodInitialRoute extends _i23.PageRouteInfo<void> {
+  const OfferFoodInitialRoute({List<_i23.PageRouteInfo>? children})
       : super(OfferFoodInitialRoute.name, initialChildren: children);
 
   static const String name = 'OfferFoodInitialRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i15.OfferFoodInitialScreen();
+      return const _i16.OfferFoodInitialScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i17.OfferFoodOverviewScreen]
+/// [_i18.OfferFoodOverviewScreen]
 class OfferFoodOverviewRoute
-    extends _i22.PageRouteInfo<OfferFoodOverviewRouteArgs> {
+    extends _i23.PageRouteInfo<OfferFoodOverviewRouteArgs> {
   OfferFoodOverviewRoute({
-    _i23.Key? key,
-    required List<_i28.FoodInfo> initialFoodInfos,
-    List<_i22.PageRouteInfo>? children,
+    _i24.Key? key,
+    required List<_i30.FoodInfo> initialFoodInfos,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           OfferFoodOverviewRoute.name,
           args: OfferFoodOverviewRouteArgs(
@@ -661,11 +731,11 @@ class OfferFoodOverviewRoute
 
   static const String name = 'OfferFoodOverviewRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodOverviewRouteArgs>();
-      return _i17.OfferFoodOverviewScreen(
+      return _i18.OfferFoodOverviewScreen(
         key: args.key,
         initialFoodInfos: args.initialFoodInfos,
       );
@@ -676,9 +746,9 @@ class OfferFoodOverviewRoute
 class OfferFoodOverviewRouteArgs {
   const OfferFoodOverviewRouteArgs({this.key, required this.initialFoodInfos});
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
-  final List<_i28.FoodInfo> initialFoodInfos;
+  final List<_i30.FoodInfo> initialFoodInfos;
 
   @override
   String toString() {
@@ -690,7 +760,7 @@ class OfferFoodOverviewRouteArgs {
     if (identical(this, other)) return true;
     if (other is! OfferFoodOverviewRouteArgs) return false;
     return key == other.key &&
-        const _i25.ListEquality().equals(
+        const _i26.ListEquality().equals(
           initialFoodInfos,
           other.initialFoodInfos,
         );
@@ -698,49 +768,49 @@ class OfferFoodOverviewRouteArgs {
 
   @override
   int get hashCode =>
-      key.hashCode ^ const _i25.ListEquality().hash(initialFoodInfos);
+      key.hashCode ^ const _i26.ListEquality().hash(initialFoodInfos);
 }
 
 /// generated route for
-/// [_i18.OrderShippingOfBoxesScreen]
-class OrderShippingOfBoxesRoute extends _i22.PageRouteInfo<void> {
-  const OrderShippingOfBoxesRoute({List<_i22.PageRouteInfo>? children})
+/// [_i19.OrderShippingOfBoxesScreen]
+class OrderShippingOfBoxesRoute extends _i23.PageRouteInfo<void> {
+  const OrderShippingOfBoxesRoute({List<_i23.PageRouteInfo>? children})
       : super(OrderShippingOfBoxesRoute.name, initialChildren: children);
 
   static const String name = 'OrderShippingOfBoxesRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i18.OrderShippingOfBoxesScreen();
+      return const _i19.OrderShippingOfBoxesScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i19.ProfileScreen]
-class ProfileRoute extends _i22.PageRouteInfo<void> {
-  const ProfileRoute({List<_i22.PageRouteInfo>? children})
+/// [_i20.ProfileScreen]
+class ProfileRoute extends _i23.PageRouteInfo<void> {
+  const ProfileRoute({List<_i23.PageRouteInfo>? children})
       : super(ProfileRoute.name, initialChildren: children);
 
   static const String name = 'ProfileRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i19.ProfileScreen();
+      return const _i20.ProfileScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i20.ThankYouScreen]
-class ThankYouRoute extends _i22.PageRouteInfo<ThankYouRouteArgs> {
+/// [_i21.ThankYouScreen]
+class ThankYouRoute extends _i23.PageRouteInfo<ThankYouRouteArgs> {
   ThankYouRoute({
-    _i23.Key? key,
+    _i24.Key? key,
     required bool isSuccess,
     required String message,
-    List<_i22.PageRouteInfo>? children,
+    List<_i23.PageRouteInfo>? children,
   }) : super(
           ThankYouRoute.name,
           args: ThankYouRouteArgs(
@@ -753,11 +823,11 @@ class ThankYouRoute extends _i22.PageRouteInfo<ThankYouRouteArgs> {
 
   static const String name = 'ThankYouRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<ThankYouRouteArgs>();
-      return _i20.ThankYouScreen(
+      return _i21.ThankYouScreen(
         key: args.key,
         isSuccess: args.isSuccess,
         message: args.message,
@@ -773,7 +843,7 @@ class ThankYouRouteArgs {
     required this.message,
   });
 
-  final _i23.Key? key;
+  final _i24.Key? key;
 
   final bool isSuccess;
 
@@ -798,17 +868,17 @@ class ThankYouRouteArgs {
 }
 
 /// generated route for
-/// [_i21.WhatsNewScreen]
-class WhatsNewRoute extends _i22.PageRouteInfo<void> {
-  const WhatsNewRoute({List<_i22.PageRouteInfo>? children})
+/// [_i22.WhatsNewScreen]
+class WhatsNewRoute extends _i23.PageRouteInfo<void> {
+  const WhatsNewRoute({List<_i23.PageRouteInfo>? children})
       : super(WhatsNewRoute.name, initialChildren: children);
 
   static const String name = 'WhatsNewRoute';
 
-  static _i22.PageInfo page = _i22.PageInfo(
+  static _i23.PageInfo page = _i23.PageInfo(
     name,
     builder: (data) {
-      return const _i21.WhatsNewScreen();
+      return const _i22.WhatsNewScreen();
     },
   );
 }