+++ lib/features/food/presentation/widget/checkup/food_boxes_checkup_tile_verified.dart
@@ -1,11 +1,21 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:zachranobed/common/presentation/utils/build_context_extensions.dart';
+import 'package:zachranobed/common/presentation/widget/button/ui_icon_button.dart';
+import 'package:zachranobed/common/presentation/widget/button/ui_primary_button.dart';
 import 'package:zachranobed/common/presentation/widget/card/ui_notification_tile.dart';
+import 'package:zachranobed/common/presentation/widget/graphics/ui_icon.dart';
+import 'package:zachranobed/common/presentation/widget/overlay/ui_dialog.dart';
 
 /// A badge widget displayed when food boxes have been verified.
 class FoodBoxesCheckupTileVerified extends StatelessWidget {
+  /// Whether the verification was performed by the user. The info button is
+  /// only shown in that case, since the dialog content speaks to the user's
+  /// own action.
+  final bool isVerifiedByUser;
+
   /// Creates a [FoodBoxesCheckupTileVerified] widget.
-  const FoodBoxesCheckupTileVerified({super.key});
+  const FoodBoxesCheckupTileVerified({super.key, required this.isVerifiedByUser});
 
   @override
   Widget build(BuildContext context) {
@@ -13,6 +23,33 @@ class FoodBoxesCheckupTileVerified extends StatelessWidget {
       title: context.l10n.foodBoxesCheckupVerifiedLabel,
       icon: Icons.check_rounded,
       iconColor: context.uiColors.success,
+      trailing: isVerifiedByUser
+          ? UiIconButton.gradient(
+              icon: Icons.info_outline,
+              onPressed: () => _showVerifiedInfo(context),
+            )
+          : null,
+    );
+  }
+
+  void _showVerifiedInfo(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (context) => UiDialog(
+        title: context.l10n.foodBoxesCheckupVerifiedTitle,
+        content: context.l10n.foodBoxesCheckupVerifiedDescription,
+        titleAlign: TextAlign.center,
+        icon: UiIcon(
+          spec: UiIconSpec.data(Icons.info_outline),
+          color: context.uiColors.textPrimary,
+        ),
+        actions: [
+          UiPrimaryButton(
+            text: context.l10n.commonClose,
+            onPressed: () => context.router.maybePop(),
+          ),
+        ],
+      ),
     );
   }
 }